KurottoUninformed
My first true venture in AI. Kurotto is a grid puzzle game similar in form and rules to sudoku.  Within a grid, the number in a circle indicates the number of continuous black square extending from a cell vertically or horizontally. Empty circles can have any number of black cells extending from them. Cells with circles can't be colored. For a more in depth explanation, see here: https://www.gmpuzzles.com/blog/kurotto-rules-and-info/

This game was used as a basis for this project demonstrating an uninformed search of a Kurotto board where in the agent goes through thousands of possible board states looking for a goal board state. I did this using a binary search. Python's more functional aspects, especially with processing lists and list comprehensions, became very helpful when sorting the results.